Fix TF cumulus module to support non-private API Gateway (#1499)
* Fix TF cumulus module to support non-private API Gateway

Fixes deployment error caused by setting private_archive_api_gateway
to false by allowing vpc_id to be null and `lambda_subnet_ids` to be an
empty list ([]).  The `subnet_ids` var is replaced by the 2 vars
`lambda_subnet_ids` and `ecs_cluster_instance_subnet_ids`, which now
surfaces this distinction that already exists within the `cumulus`
module.

This still requires the policy for the API Gateway to be manually
edited or removed, if necessary, because the policy will not be
updated automatically due to NGAP requirements.